Implement room rendering layout with tileset

+++ GLFWProject/texture.cpp
@@ -54,7 +54,7 @@ void Texture::bind()
 	glBindTexture(GL_TEXTURE_2D, m_id);
 }
 
-glm::vec2 Texture::getSize() const
+glm::ivec2 Texture::getSize() const
 {
-	return glm::vec2(m_width, m_height);
+	return glm::ivec2(m_width, m_height);
 }
